LPS-89943 Do not allow to delete the production collection

+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/internal/CTEngineManagerImpl.java
@@ -124,6 +124,17 @@ public class CTEngineManagerImpl implements CTEngineManager {
 			return;
 		}
 
+		CTCollection ctCollection = ctCollectionOptional.get();
+
+		if (ctCollection.isProduction()) {
+			if (_log.isWarnEnabled()) {
+				_log.warn(
+					"Deleting the production change collection is forbidden");
+			}
+
+			return;
+		}
+
 		try {
 			_ctCollectionLocalService.deleteCTCollection(
 				ctCollectionOptional.get());
